Add multi users buy order function

+++ app/src/main/java/com/mingmin/sharebuy/Buyer.java
@@ -36,7 +36,7 @@ public class Buyer {
         return orderTime;
     }
 
-    public long getBuyCount() {
+    public int getBuyCount() {
         return buyCount;
     }
